emacs: improved my tidal setup function
Previously the tidal setup function, the one that started jack,
sclang, haskell and opened a file, did all this automatically,
all at the same time. Some of those must be instead started one
after the other, so I opted for a procedure that prompts me with
step shortcuts, that allow me to direct the setup process.